# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/categories/{id}/attributes/{attributeId}": {
      "delete": {
        "operationId": "CategoryController_removeAttribute",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "attributeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "მახასიათებელი მოიხსნა"
          },
          "404": {
            "description": "ეს მახასიათებელი პირდაპირ მიბმული არ არის ამ კატეგორიაზე"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "მახასიათებლის მოხსნა კატეგორიიდან (ADMIN)",
        "tags": [
          "categories"
        ]
      }
    },
...
    "/attributes": {
      "get": {
        "operationId": "AttributeController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "გვერდის ნომერი (იწყება 1-დან)",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "ჩანაწერების რაოდენობა თითო გვერდზე",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "დალაგების ველი",
            "schema": {
              "default": "createdAt",
              "example": "createdAt",
              "type": "string"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "description": "დალაგების მიმართულება",
            "schema": {
              "default": "DESC",
              "example": "DESC",
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ]
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "description": "მხოლოდ ამ ტიპის attribute-ები",
            "schema": {
              "type": "string",
              "enum": [
                "select",
                "multi_select",
                "number",
                "text",
                "boolean",
                "range"
              ]
            }
          },
          {
            "name": "isFilterable",
            "required": false,
            "in": "query",
            "description": "მხოლოდ filterable (ან non-filterable) attribute-ები",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "მახასიათებლების გვერდიანი სია"
          }
        },
        "summary": "მახასიათებლების გვერდიანი სია",
        "tags": [
          "attributes"
        ]
      },
      "post": {
        "operationId": "AttributeController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAttributeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "მახასიათებელი შეიქმნა"
          },
          "400": {
            "description": "ვალიდაციის შეცდომა"
          },
          "409": {
            "description": "code უკვე დაკავებულია"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "ახალი მახასიათებლის შექმნა (ADMIN)",
        "tags": [
          "attributes"
        ]
      }
    },
    "/attributes/{id}": {
      "get": {
        "operationId": "AttributeController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "მახასიათებელი"
          },
          "404": {
            "description": "მახასიათებელი ვერ მოიძებნა"
          }
        },
        "summary": "კონკრეტული მახასიათებლის მიღება (options-ითურთ)",
        "tags": [
          "attributes"
        ]
      },
      "put": {
        "operationId": "AttributeController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAttributeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "მახასიათებელი განახლდა"
          },
          "404": {
            "description": "მახასიათებელი ვერ მოიძებნა"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "მახასიათებლის განახლება (ADMIN)",
        "tags": [
          "attributes"
        ]
      },
      "delete": {
        "operationId": "AttributeController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "მახასიათებელი წაიშალა"
          },
          "404": {
            "description": "მახასიათებელი ვერ მოიძებნა"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "მახასიათებლის წაშლა (ADMIN)",
        "tags": [
          "attributes"
        ]
      }
    },
    "/attributes/{id}/options": {
      "post": {
        "operationId": "AttributeController_addOption",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAttributeOptionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "ოფცია დაემატა"
          },
          "400": {
            "description": "ოფცია დასაშვებია მხოლოდ select/multi_select ტიპებზე"
          },
          "409": {
            "description": "code უკვე დაკავებულია ამ მახასიათებელზე"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "მახასიათებელზე ახალი ოფციის დამატება (ADMIN)",
        "tags": [
          "attributes"
        ]
      }
    },
    "/attributes/{id}/options/{optionId}": {
      "put": {
        "operationId": "AttributeController_updateOption",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "optionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAttributeOptionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "ოფცია განახლდა"
          },
          "404": {
            "description": "ოფცია ვერ მოიძებნა"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "ოფციის განახლება (ADMIN)",
        "tags": [
          "attributes"
        ]
      },
      "delete": {
        "operationId": "AttributeController_removeOption",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "optionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ოფცია წაიშალა"
          },
          "404": {
            "description": "ოფცია ვერ მოიძებნა"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "ოფციის წაშლა (ADMIN)",
        "tags": [
          "attributes"
        ]
      }
    },
...
  },
  "info": {
    "title": "Online Shop API",
    "description": "REST API for the online shop backend",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "NameTranslationDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "სახელი",
            "example": "აკუმულატორები"
          }
        },
        "required": [
          "name"
        ]
      },
      "NameTranslationsDto": {
        "type": "object",
        "properties": {
          "ka": {
            "$ref": "#/components/schemas/NameTranslationDto"
          },
          "en": {
            "$ref": "#/components/schemas/NameTranslationDto"
          },
          "ru": {
            "$ref": "#/components/schemas/NameTranslationDto"
          }
        },
        "required": [
          "ka"
        ]
      },
...
      "CreateAttributeDto": {
        "type": "object",
        "properties": {
          "translations": {
            "description": "მრავალენოვანი სახელი — { ka: {name}, en?, ru? }, ka სავალდებულოა",
            "allOf": [
              {
                "$ref": "#/components/schemas/NameTranslationsDto"
              }
            ]
          },
          "code": {
            "type": "string",
            "pattern": "^[a-z0-9]+(-[a-z0-9]+)*$",
            "description": "უნიკალური კოდი (frontend filter query-ებისთვის), მხოლოდ ლათინური ასოები/ციფრები/ტირე",
            "example": "amperage"
          },
          "type": {
            "enum": [
              "select",
              "multi_select",
              "number",
              "text",
              "boolean",
              "range"
            ],
            "type": "string",
            "description": "მახასიათებლის ტიპი — განსაზღვრავს, product_attribute_value-ში რომელი value ველი გამოიყენება",
            "example": "select"
          },
          "unit": {
            "type": "string",
            "description": "ერთეული (Ah, V, mm...) — number/range ტიპებისთვის",
            "example": "Ah"
          },
          "isFilterable": {
            "type": "boolean",
            "description": "გამოჩნდება თუ არა filter-ებში",
            "default": true
          },
          "isRequired": {
            "type": "boolean",
            "description": "სავალდებულოა თუ არა admin ფორმაში",
            "default": false
          },
          "sortOrder": {
            "type": "number",
            "description": "დალაგების რიგი",
            "example": 0,
            "default": 0
          }
        },
        "required": [
          "translations",
          "code",
          "type"
        ]
      },
      "UpdateAttributeDto": {
        "type": "object",
        "properties": {
          "translations": {
            "description": "მრავალენოვანი სახელი — { ka: {name}, en?, ru? }, ka სავალდებულოა",
            "allOf": [
              {
                "$ref": "#/components/schemas/NameTranslationsDto"
              }
            ]
          },
          "code": {
            "type": "string",
            "pattern": "^[a-z0-9]+(-[a-z0-9]+)*$",
            "description": "უნიკალური კოდი (frontend filter query-ებისთვის), მხოლოდ ლათინური ასოები/ციფრები/ტირე",
            "example": "amperage"
          },
          "type": {
            "enum": [
              "select",
              "multi_select",
              "number",
              "text",
              "boolean",
              "range"
            ],
            "type": "string",
            "description": "მახასიათებლის ტიპი — განსაზღვრავს, product_attribute_value-ში რომელი value ველი გამოიყენება",
            "example": "select"
          },
          "unit": {
            "type": "string",
            "description": "ერთეული (Ah, V, mm...) — number/range ტიპებისთვის",
            "example": "Ah"
          },
          "isFilterable": {
            "type": "boolean",
            "description": "გამოჩნდება თუ არა filter-ებში",
            "default": true
          },
          "isRequired": {
            "type": "boolean",
            "description": "სავალდებულოა თუ არა admin ფორმაში",
            "default": false
          },
          "sortOrder": {
            "type": "number",
            "description": "დალაგების რიგი",
            "example": 0,
            "default": 0
          }
        }
      },
      "ValueTranslationDto": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string",
            "description": "მნიშვნელობა",
            "example": "ბანერი"
          }
        },
        "required": [
          "value"
        ]
      },
      "ValueTranslationsDto": {
        "type": "object",
        "properties": {
          "ka": {
            "$ref": "#/components/schemas/ValueTranslationDto"
          },
          "en": {
            "$ref": "#/components/schemas/ValueTranslationDto"
          },
          "ru": {
            "$ref": "#/components/schemas/ValueTranslationDto"
          }
        },
        "required": [
          "ka"
        ]
      },
      "CreateAttributeOptionDto": {
        "type": "object",
        "properties": {
          "translations": {
            "description": "მრავალენოვანი მნიშვნელობა — { ka: {value}, en?, ru? }, ka სავალდებულოა",
            "allOf": [
              {
                "$ref": "#/components/schemas/ValueTranslationsDto"
              }
            ]
          },
          "code": {
            "type": "string",
            "pattern": "^[a-z0-9]+(-[a-z0-9]+)*$",
            "description": "უნიკალური კოდი (attribute-ის ფარგლებში), მხოლოდ ლათინური ასოები/ციფრები/ტირე",
            "example": "banner"
          },
          "sortOrder": {
            "type": "number",
            "description": "დალაგების რიგი",
            "example": 0,
            "default": 0
          }
        },
        "required": [
          "translations",
          "code"
        ]
      },
      "UpdateAttributeOptionDto": {
        "type": "object",
        "properties": {
          "translations": {
            "description": "მრავალენოვანი მნიშვნელობა — { ka: {value}, en?, ru? }, ka სავალდებულოა",
            "allOf": [
              {
                "$ref": "#/components/schemas/ValueTranslationsDto"
              }
            ]
          },
          "code": {
            "type": "string",
            "pattern": "^[a-z0-9]+(-[a-z0-9]+)*$",
            "description": "უნიკალური კოდი (attribute-ის ფარგლებში), მხოლოდ ლათინური ასოები/ციფრები/ტირე",
            "example": "banner"
          },
          "sortOrder": {
            "type": "number",
            "description": "დალაგების რიგი",
            "example": 0,
            "default": 0
          }
        }
      },
...
    }
  }
}